Move controllers and views to addon and put facades in app
