+++ addon/controllers/google-map/circles.js
@@ -1,0 +1,12 @@
+import Ember from 'ember';
+
+var computed = Ember.computed;
+
+/**
+ * @class GoogleMapCirclesController
+ * @extends Ember.ArrayController
+ */
+export default Ember.ArrayController.extend({
+  itemController: computed.alias('parentController.circleController'),
+  model:          computed.alias('parentController.circles')
+});
