+++ addon/controllers/google-map/markers.js
@@ -1,0 +1,12 @@
+import Ember from 'ember';
+
+var computed = Ember.computed;
+
+/**
+ * @class GoogleMapMarkersController
+ * @extends Ember.ArrayController
+ */
+export default Ember.ArrayController.extend({
+  itemController: computed.alias('parentController.markerController'),
+  model:          computed.alias('parentController.markers')
+});
