+++ app/controllers/google-map/circles.js
@@ -1,12 +1,2 @@
-import Ember from 'ember';
-
-var computed = Ember.computed;
-
-/**
- * @class GoogleMapCirclesController
- * @extends Ember.ArrayController
- */
-export default Ember.ArrayController.extend({
-  itemController: computed.alias('parentController.circleController'),
-  model:          computed.alias('parentController.circles')
-});
+import GoogleMapCirclesController from 'ember-google-map/controllers/google-map/circles';
+export default GoogleMapCirclesController;
