+++ app/controllers/google-map/info-window.js
@@ -1,18 +1,2 @@
-import Ember from 'ember';
-
-var computed = Ember.computed;
-var alias = computed.alias;
-
-/**
- * @class GoogleMapInfoWindowController
- * @extends Ember.Controller
- */
-export default Ember.Controller.extend({
-  templateName: alias('model.templateName'),
-  zIndex:       alias('model.zIndex'),
-  lat:          alias('model.lat'),
-  lng:          alias('model.lng'),
-  isVisible:    alias('model.isVisible'),
-  title:        alias('model.title'),
-  description:  alias('model.description')
-});
+import GoogleMapInfoWindowController from 'ember-google-map/controllers/google-map/info-window';
+export default GoogleMapInfoWindowController;
