+++ app/controllers/google-map/info-windows.js
@@ -1,12 +1,2 @@
-import Ember from 'ember';
-
-var computed = Ember.computed;
-
-/**
- * @class GoogleMapInfoWindowsController
- * @extends Ember.ArrayController
- */
-export default Ember.ArrayController.extend({
-  itemController: computed.alias('parentController.infoWindowController'),
-  model:          computed.alias('parentController.infoWindows')
-});
+import GoogleMapInfoWindowsController from 'ember-google-map/controllers/google-map/info-windows';
+export default GoogleMapInfoWindowsController;
