+++ app/controllers/google-map/markers.js
@@ -1,12 +1,2 @@
-import Ember from 'ember';
-
-var computed = Ember.computed;
-
-/**
- * @class GoogleMapMarkersController
- * @extends Ember.ArrayController
- */
-export default Ember.ArrayController.extend({
-  itemController: computed.alias('parentController.markerController'),
-  model:          computed.alias('parentController.markers')
-});
+import GoogleMapMarkersController from 'ember-google-map/controllers/google-map/markers';
+export default GoogleMapMarkersController;
