+++ app/controllers/google-map/polygons.js
@@ -1,14 +1,2 @@
-import Ember from 'ember';
-import GoogleMapPolylinesController from './polylines';
-
-var computed = Ember.computed;
-
-/**
- * @class GoogleMapPolygonsController
- * @extends GoogleMapPolylinesController
- */
-export default GoogleMapPolylinesController.extend({
-  itemController: computed.alias('parentController.polygonController'),
-  model:          computed.alias('parentController.polygons'),
-  pathController: computed.alias('parentController.polygonPathController')
-});
+import GoogleMapPolygonsController from 'ember-google-map/controllers/google-map/polygons';
+export default GoogleMapPolygonsController;
