+++ app/controllers/google-map/polylines.js
@@ -1,13 +1,2 @@
-import Ember from 'ember';
-
-var computed = Ember.computed;
-
-/**
- * @class GoogleMapPolylinesController
- * @extends Ember.ArrayController
- */
-export default Ember.ArrayController.extend({
-  itemController: computed.alias('parentController.polylineController'),
-  model:          computed.alias('parentController.polylines'),
-  pathController: computed.alias('parentController.polylinePathController')
-});
+import GoogleMapPolylinesController from 'ember-google-map/controllers/google-map/polylines';
+export default GoogleMapPolylinesController;
